chore: decouple services factory from core
